// packages/flavor-healthfit/src/integrations/hevy/generated/hevy-api.generated.ts
export namespace Schemas {
// ...
  export type Workout = Partial<{
    id: string;
    title: string;
    routine_id: string;
    description: string;
    start_time: string;
    end_time: string;
    updated_at: string;
    created_at: string;
    exercises: Array<
      Partial<{
        index: number;
        title: string;
        notes: string;
        exercise_template_id: string;
        supersets_id: number | null;
        sets: Array<
          Partial<{
            index: number;
            type: string;
            weight_kg: number | null;
            reps: number | null;
            distance_meters: number | null;
            duration_seconds: number | null;
            rpe: number | null;
            custom_metric: number | null;
          }>
        >;
      }>
    >;
  }>;
// ...
}